Add Cache for Images
In order to not extract every single image from the ZIP/Rar every time the whole archive will be unpacked.
In addition every image will be resized if it is larger than predefined max values

This whole process starts when the first image is loaded. And can take several seconds. Followup commits will change this.